Detect problems with the parsing of date/times

any_problems() takes a list of capture matrices (see parse_dttm()) and reports on parsing problems by means of predicate values. A FALSE value indicates that the parsing was successful and a TRUE value a parsing failure in at least one of the inputs to create_iso8601(). Note that this is an internal function to be used in the context of create_iso8601() source code and hence each capture matrix corresponds to one input to create_iso8601().

Usage

any_problems(cap_matrices, .cutoff_2000 = 68L)

Arguments

cap_matrices

A list of capture matrices in the sense of the returned value by parse_dttm().

.cutoff_2000

An integer value. Two-digit years smaller or equal to .cutoff_2000 are parsed as though starting with 20, otherwise parsed as though starting with 19.

Value

A logical whose length matches the number of underlying date/times passed as inputs to create_iso8601(), i.e. whose length matches the number of rows of the capture matrices in cap_matrices.

Examples

# No problem (return value is `FALSE`).
sdtm.oak:::any_problems(list(sdtm.oak:::parse_dttm("1980-06-18", "y-m-d")))
#> [1] FALSE

# Now the parsing fails (return value is `TRUE`).
sdtm.oak:::any_problems(list(sdtm.oak:::parse_dttm("1980-06-18", "ymd")))
#> [1] TRUE

# Find if there has been a problem in either in the `date` or `time` inputs.
# The following problems are expected with:
#   - `"2001/12/25"` as it won't be parsed with the format `"y-m-d"`
#   - `"00h12m21"` as it won't be parsed with the format `"H:M:S"`.
#
date <- c("2000-01-05", "2001/12/25", "1980-06-18", "1979-09-07")
time <- c("00h12m21", "22:35:05", "03:00:15", "07:09:00")

cap_matrix_date <- sdtm.oak:::parse_dttm(date, "y-m-d")
cap_matrix_time <- sdtm.oak:::parse_dttm(time, "H:M:S")

(cap_matrices <- list(cap_matrix_date, cap_matrix_time))
#> [[1]]
#>      year   mon  mday hour min sec
#> [1,] "2000" "01" "05" NA   NA  NA 
#> [2,] NA     NA   NA   NA   NA  NA 
#> [3,] "1980" "06" "18" NA   NA  NA 
#> [4,] "1979" "09" "07" NA   NA  NA 
#> 
#> [[2]]
#>      year mon mday hour min  sec 
#> [1,] NA   NA  NA   NA   NA   NA  
#> [2,] NA   NA  NA   "22" "35" "05"
#> [3,] NA   NA  NA   "03" "00" "15"
#> [4,] NA   NA  NA   "07" "09" "00"
#> 

# `any_problems()` returns `TRUE` for the first two elements because of the
# failure to parse `"2001/12/25"` and `"00h12m21"`, respectively.
sdtm.oak:::any_problems(cap_matrices)
#> [1]  TRUE  TRUE FALSE FALSE