Coalesce capture matrices

coalesce_capture_matrices() combines several capture matrices into one. Each argument of ... should be a capture matrix in the sense of the output by complete_capture_matrix(), meaning a character matrix of six columns whose names are: year, mon, mday, hour, min or sec.

Usage

coalesce_capture_matrices(...)

Arguments

...

A sequence of capture matrices.

Value

A single capture matrix whose values have been coalesced in the sense of coalesce().

Examples

cols <- c("year", "mon", "mday", "hour", "min", "sec")
dates <- c("2020", "01", "01", "20", NA, NA)
times <- c(NA, NA, NA, "10", "00", "05")
m_dates <- matrix(dates, nrow = 1L, ncol = 6L, dimnames = list(NULL, cols))
m_times <- matrix(times, nrow = 1L, ncol = 6L, dimnames = list(NULL, cols))

# Note how the hour "20" takes precedence over "10"
sdtm.oak:::coalesce_capture_matrices(m_dates, m_times)
#>      year   mon  mday hour min  sec 
#> [1,] "2020" "01" "01" "20" "00" "05"

# Reverse the order of the inputs and now hour "10" takes precedence
sdtm.oak:::coalesce_capture_matrices(m_times, m_dates)
#>      year   mon  mday hour min  sec 
#> [1,] "2020" "01" "01" "10" "00" "05"

# Single inputs should result in the same output as the input
sdtm.oak:::coalesce_capture_matrices(m_dates)
#>      year   mon  mday hour min sec
#> [1,] "2020" "01" "01" "20" NA  NA 
sdtm.oak:::coalesce_capture_matrices(m_times)
#>      year mon mday hour min  sec 
#> [1,] NA   NA  NA   "10" "00" "05"