Controlled terminology mappings

ct_mappings() takes a controlled terminology specification and returns the mappings in the form of a tibble in long format, i.e. the recoding of values in the from column to the to column values, one mapping per row.

The resulting mappings are unique, i.e. if from values are duplicated in two from columns, the first column indicated in from takes precedence, and only that mapping is retained in the controlled terminology map.

Usage

ct_mappings(ct_spec, from = ct_spec_vars("from"), to = ct_spec_vars("to"))

Arguments

ct_spec

Controlled terminology specification as a tibble. Each row is for a mapped controlled term. Controlled terms are expected in the column indicated by to_col.

from

A character vector of column names indicating the variables containing values to be recoded.

to

A single string indicating the column whose values are to be recoded into.

Value

A tibble with two columns, from and to, indicating the mapping of values, one per row.

Examples

# Read in a bundled controlled terminology spec example (ex. 01).
(ct_spec_01 <- read_ct_spec_example("ct-01-cm"))
#> # A tibble: 33 × 8
#>    codelist_code term_code CodedData term_value collected_value    
#>    <chr>         <chr>     <chr>     <chr>      <chr>              
#>  1 C71113        C25473    QD        QD         QD (Every Day)     
#>  2 C71113        C64496    BID       BID        BID (Twice a Day)  
#>  3 C71113        C64499    PRN       PRN        PRN (As Needed)    
#>  4 C71113        C64516    Q2H       Q2H        Q2H (Every 2 Hours)
#>  5 C71113        C64530    QID       QID        QID (4 Times a Day)
#>  6 C66726        C25158    CAPSULE   CAPSULE    Capsule            
#>  7 C66726        C25394    PILL      PILL       Pill               
#>  8 C66726        C29167    LOTION    LOTION     Lotion             
#>  9 C66726        C42887    AEROSOL   AEROSOL    Aerosol            
#> 10 C66726        C42944    INHALANT  INHALANT   Inhalant           
#> # ℹ 23 more rows
#> # ℹ 3 more variables: term_preferred_term <chr>, term_synonyms <chr>,
#> #   raw_codelist <chr>

# Generate mappings from the terminology specification.
sdtm.oak:::ct_mappings(ct_spec = ct_spec_01)
#> # A tibble: 79 × 2
#>    from     to      
#>    <chr>    <chr>   
#>  1 QD       QD      
#>  2 BID      BID     
#>  3 PRN      PRN     
#>  4 Q2H      Q2H     
#>  5 QID      QID     
#>  6 CAPSULE  CAPSULE 
#>  7 PILL     PILL    
#>  8 LOTION   LOTION  
#>  9 AEROSOL  AEROSOL 
#> 10 INHALANT INHALANT
#> # ℹ 69 more rows

# Take a glimpse at those mappings where an actual recoding happens.
sdtm.oak:::ct_mappings(ct_spec = ct_spec_01) |>
  dplyr::filter(from != to) |>
  print(n = 20)
#> # A tibble: 49 × 2
#>    from                 to             
#>    <chr>                <chr>          
#>  1 QD (Every Day)       QD             
#>  2 BID (Twice a Day)    BID            
#>  3 PRN (As Needed)      PRN            
#>  4 Q2H (Every 2 Hours)  Q2H            
#>  5 QID (4 Times a Day)  QID            
#>  6 Capsule              CAPSULE        
#>  7 Pill                 PILL           
#>  8 Lotion               LOTION         
#>  9 Aerosol              AEROSOL        
#> 10 Inhalant             INHALANT       
#> 11 Injection            INJECTION      
#> 12 Liquid               LIQUID         
#> 13 Tablet               TABLET         
#> 14 Yes                  Y              
#> 15 IM (Intramuscular)   INTRAMUSCULAR  
#> 16 EP (Epidural)        EPIDURAL       
#> 17 IA (Intra-arterial)  INTRA-ARTERIAL 
#> 18 IJ (Intra-articular) INTRA-ARTICULAR
#> 19 OP (Ophthalmic)      OPHTHALMIC     
#> 20 PO (Oral)            ORAL           
#> # ℹ 29 more rows