Find the path to an example SDTM domain file

domain_example() resolves the local path to a SDTM domain example file. The domain examples files were imported from pharmaversesdtm. See Details section for available datasets.

Usage

domain_example(example)

Source

See https://cran.r-project.org/package=pharmaversesdtm.

Arguments

example

A string with either the basename, file name, or relative path to a SDTM domain example file bundled with {stdm.oak}, e.g. "cm" (Concomitant Medication) or "ae" (Adverse Events).

Value

The local path to an example file if example is supplied, or a character vector of example file names.

Details

Datasets were obtained from pharmaversesdtm but are originally sourced from the CDISC pilot project or have been constructed ad-hoc by the admiral team. These datasets are bundled with {sdtm.oak}, thus obviating a dependence on {pharmaversesdtm}.

Example SDTM domains

"ae": Adverse Events (AE) data set.

"apsc": Associated Persons Subject Characteristics (APSC) data set.

"cm": Concomitant Medications (CM) data set.

"vs": Vital Signs (VS) data set.

See also

Examples

# If no example is provided it returns a vector of possible choices.
domain_example()
#> [1] "ae"   "apsc" "cm"   "vs"  

# Get the local path to the Concomitant Medication dataset file.
domain_example("cm")
#> [1] "/renv/lib/R-4.3/x86_64-pc-linux-gnu/sdtm.oak/domain/cm.rds"

# Local path to the Adverse Events dataset file.
domain_example("ae")
#> [1] "/renv/lib/R-4.3/x86_64-pc-linux-gnu/sdtm.oak/domain/ae.rds"