Parse a date/time format

parse_dttm_fmt() parses a date/time formats, meaning it will try to parse the components of the format fmt that refer to date/time components. parse_dttm_fmt_() is similar to parse_dttm_fmt() but is not vectorized over fmt.

Usage

parse_dttm_fmt_(fmt, pattern)

parse_dttm_fmt(fmt, patterns = fmt_cmp())

Arguments

fmt

A format string (scalar) to be parsed by patterns.

pattern, patterns

A string (in the case of pattern), or a character vector (in the case of patterns) of regexps for each of the individual date/time components. Default value is that of fmt_cmp(). Use this function if you plan on passing a different set of patterns.

Value

A tibble of seven columns:

fmt_c: date/time format component. Values are either "year", "mon", "mday", "hour", "min", "sec", or NA.

pat: Regexp used to parse the date/time component.

cap: The captured substring from the format.

start: Start position in the format string for this capture.

end: End position in the format string for this capture.

len: Length of the capture (number of chars).

ord: Ordinal of this date/time component in the format string.

Each row is for either a date/time format component or a "delimiter" string or pattern in-between format components.

Examples

sdtm.oak:::parse_dttm_fmt("ymd")
#> # A tibble: 3 × 7
#>   fmt_c pat   cap   start   end   len   ord
#>   <chr> <chr> <chr> <int> <int> <int> <int>
#> 1 year  y+    y         1     1     1     1
#> 2 mon   m+    m         2     2     1     2
#> 3 mday  d+    d         3     3     1     3
sdtm.oak:::parse_dttm_fmt("H:M:S")
#> # A tibble: 5 × 7
#>   fmt_c pat   cap   start   end   len   ord
#>   <chr> <chr> <chr> <int> <int> <int> <int>
#> 1 hour  H+    H         1     1     1     1
#> 2 NA    NA    :         2     2     1    NA
#> 3 min   M+    M         3     3     1     2
#> 4 NA    NA    :         4     4     1    NA
#> 5 sec   S+    S         5     5     1     3

sdtm.oak:::parse_dttm_fmt("ymd HMS")
#> # A tibble: 7 × 7
#>   fmt_c pat   cap   start   end   len   ord
#>   <chr> <chr> <chr> <int> <int> <int> <int>
#> 1 year  y+    "y"       1     1     1     1
#> 2 mon   m+    "m"       2     2     1     2
#> 3 mday  d+    "d"       3     3     1     3
#> 4 NA    NA    " "       4     4     1    NA
#> 5 hour  H+    "H"       5     5     1     4
#> 6 min   M+    "M"       6     6     1     5
#> 7 sec   S+    "S"       7     7     1     6

# Repeating the same special patterns, e.g. "yy" still counts as one pattern
# only.
sdtm.oak:::parse_dttm_fmt("yymmdd HHMMSS")
#> # A tibble: 7 × 7
#>   fmt_c pat   cap   start   end   len   ord
#>   <chr> <chr> <chr> <int> <int> <int> <int>
#> 1 year  y+    "yy"      1     2     2     1
#> 2 mon   m+    "mm"      3     4     2     2
#> 3 mday  d+    "dd"      5     6     2     3
#> 4 NA    NA    " "       7     7     1    NA
#> 5 hour  H+    "HH"      8     9     2     4
#> 6 min   M+    "MM"     10    11     2     5
#> 7 sec   S+    "SS"     12    13     2     6

# Note that `"y"`, `"m"`, `"d"`, `"H"`, `"M"` or `"S"` are reserved patterns
# that are matched first and interpreted as format components. # Example: the
# first "y" in "year" is parsed as meaning year followed by "ear y". The
# second "y" is not longer matched because a first match already # succeded.
sdtm.oak:::parse_dttm_fmt("year y")
#> # A tibble: 2 × 7
#>   fmt_c pat   cap   start   end   len   ord
#>   <chr> <chr> <chr> <int> <int> <int> <int>
#> 1 year  y+    y         1     1     1     1
#> 2 NA    NA    ear y     2     6     5    NA

# Specify custom patterns
sdtm.oak:::parse_dttm_fmt(
  "year month day",
  fmt_cmp(year = "year", mon = "month", mday = "day")
)
#> # A tibble: 5 × 7
#>   fmt_c pat   cap     start   end   len   ord
#>   <chr> <chr> <chr>   <int> <int> <int> <int>
#> 1 year  year  "year"      1     4     4     1
#> 2 NA    NA    " "         5     5     1    NA
#> 3 mon   month "month"     6    10     5     2
#> 4 NA    NA    " "        11    11     1    NA
#> 5 mday  day   "day"      12    14     3     3